backend/requirements.txt
amqp==5.3.1
anyio
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
botocore
boto3
bs4==0.0.2
celery==5.4.0
certifi
cffi==1.17.1
charset-normalizer
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20241021
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.4
django-js-asset==3.0.1
django-mptt==0.16.0
django-redis==5.4.0
django-storages==1.14.4
django-unfold==0.44.0
djangorestframework==3.15.2
djangorestframework-simplejwt
drf-yasg==1.21.8
eitaa==2.1.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
idna==3.10
import-export==0.3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
minio==7.2.13
openapi-codec==1.3.2
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.2
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
sentry-sdk==2.19.2
setuptools
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlparse==0.5.3
tablib==3.7.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary
utils
jdatetime
user-agents
jwt
chromadb
# sentence-transformers
openai